requirements.txt
numpy==1.21.0
pandas==1.3.0
scikit-learn==0.24.2
scikit-surprise==1.1.1
matplotlib==3.4.2
seaborn==0.11.1